+++ Sources/WalletConnect/Types/Pairing/PairingSequenceState.swift
@@ -19,22 +19,3 @@ extension PairingType {
         }
     }
 }
-
-extension SessionType {
-    enum SequenceState: Codable, Equatable {
-        case settled(Settled)
-        case pending(Pending)
-        enum CodingKeys: CodingKey {
-            case settled
-            case pending
-        }
-        var topic: String {
-            switch self {
-            case .settled(let sequence):
-                return sequence.topic
-            case .pending(let sequence):
-                return sequence.topic
-            }
-        }
-    }
-}
